refactor: separate project generator from code generator interface
